requirements.txt
selenium
pytz
python-dateutil
pytest
pytest-mock
tweepy
mechanicalsoup
bs4
setuptools
soupsieve
oauthlib
lxml
six>=1.10.0
requests>=2.11.1
urllib3>=1.21.1,<1.25
